# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VK Bot API",
        "description": "API для управления VK ботом",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/settings/save": {
            "post": {
                "tags": [
                    "Settings"
                ],
                "summary": "Сохранить настройки приложения",
                "description": "Обновляет настройки приложения",
                "operationId": "397dbab512c95e01ed1ec22a57fe07ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "show_followers",
                                    "show_friends",
                                    "task_timeout"
                                ],
                                "properties": {
                                    "show_followers": {
                                        "description": "Отображение подписчиков",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "show_friends": {
                                        "description": "Отображение друзей",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "task_timeout": {
                                        "description": "Таймаут задач в секундах",
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Настройки успешно сохранены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Настройки успешно сохранены"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "show_followers": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "show_friends": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "task_timeout": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка валидации"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}